% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{add_pip_class}
\alias{add_pip_class}
\title{add PIP class}
\usage{
add_pip_class(x, pip_type)
}
\arguments{
\item{x}{data frame}

\item{pip_type}{character: One of "md", "id", "gd_*". Generally this comes
from the output of \code{\link[=identify_pip_type]{identify_pip_type()}}}
}
\value{
data frame with pip class
}
\description{
add PIP class
}
\keyword{internal}
