% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{as_pipgd}
\alias{as_pipgd}
\title{Add pipgd class for group data}
\usage{
as_pipgd(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame with new class pipgd
}
\description{
Add pipgd class for group data
}
