% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{as_pipid}
\alias{as_pipid}
\title{Add pipid class for imputed data}
\usage{
as_pipid(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame with new class pipid
}
\description{
Add pipid class for imputed data
}
