% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{as_pipmd}
\alias{as_pipmd}
\title{Add pipmd class for microdata}
\usage{
as_pipmd(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame with new class pipmd
}
\description{
Add pipmd class for microdata
}
