% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{assign_pipclass}
\alias{assign_pipclass}
\title{assign correct class to DLW data}
\usage{
assign_pipclass(df)
}
\arguments{
\item{df}{dataframe loaded from DLW flat structure}
}
\value{
data.table
}
\description{
assign correct class to DLW data
}
