% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_pipgd_params}
\alias{check_pipgd_params}
\title{Check parameters of pipgd functions}
\usage{
check_pipgd_params(lp)
}
\arguments{
\item{lp}{list of parameters}
}
\value{
invisible TRUE
}
\description{
Check parameters of pipgd functions
}
\keyword{internal}
