% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_pip_format.R
\name{convert_to_pip_format}
\alias{convert_to_pip_format}
\title{Convert vectors}
\usage{
convert_to_pip_format(welfare, weight, pip_type)
}
\arguments{
\item{welfare}{numeric: welfare vector}

\item{weight}{numeric: weights of population vector}

\item{pip_type}{character: One of "md", "id", "gd_*". Generally this comes
from the output of \code{\link[=identify_pip_type]{identify_pip_type()}}}
}
\value{
data.frame
}
\description{
Convert vectors
}
