% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{convert_to_pip_md}
\alias{convert_to_pip_md}
\title{Convert vectors to microdata data PIP format}
\usage{
convert_to_pip_md(
  dt,
  welfare_var,
  imputation_id_var,
  verbose = getOption("pipster.verbose")
)
}
\arguments{
\item{dt}{data.frame with welfare data}

\item{welfare_var}{character: variable name of welfare vector in dt}

\item{imputation_id_var}{character: variable name of imputation ID vector in
dt}

\item{verbose}{logical: Whether to display important messages about your data
or query}
}
\value{
tibble
}
\description{
Convert vectors to microdata data PIP format
}
\keyword{internal}
