% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipster_md.R
\name{default_warning_message}
\alias{default_warning_message}
\title{Warning message used in default methods for all pipster S3 generic functions}
\usage{
default_warning_message(generic_name, first_arg)
}
\arguments{
\item{generic_name}{character: name of generic function of default method}

\item{first_arg}{object: first argument in the generic function, the class of
which determines the S3 method.}
}
\value{
warning message
}
\description{
Warning message used in default methods for all pipster S3 generic functions
}
\keyword{internal}
