% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_gini.default}
\alias{get_gini.default}
\title{Gini coefficient S3 default}
\usage{
\method{get_gini}{default}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments}
}
\value{
list
}
\description{
This default S3 method returns an error when called
}
