% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_gini.pipster_gd}
\alias{get_gini.pipster_gd}
\title{Gini coefficient on group data}
\usage{
\method{get_gini}{pipster_gd}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to \link{pipgd_gini}}
}
\value{
list
}
\description{
This is a method for computing the gini coefficient on group data.
Requires a pipster object of class \code{pipster_gd} (grouped data object).
This is a wrapper over \link{pipgd_gini}, which should be viewed
for more detail.
}
