% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_gini.pipster_md}
\alias{get_gini.pipster_md}
\title{Gini coefficient on micro data}
\usage{
\method{get_gini}{pipster_md}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to \link{pipmd_gini}}
}
\value{
list
}
\description{
This is a method for computing the gini coefficient on micro data.
Requires a pipster object of class \code{pipster_md} (grouped micro object).
This is a wrapper over \link{pipmd_gini}, which should be viewed
for more detail.
}
