% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_gini}
\alias{get_gini}
\title{Gini coefficient}
\usage{
get_gini(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}
}
\value{
list
}
\description{
This is a generic function calculating the gini coefficient.
}
