% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_mld.pipster_md}
\alias{get_mld.pipster_md}
\title{Mean Log Deviation (MLD) for micro data}
\usage{
\method{get_mld}{pipster_md}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to \link{pipmd_mld}}
}
\value{
list
}
\description{
This is a method to compute the MLD on group data.
Requires a pipster object of class \code{pipster_md} (grouped data object).
This is a wrapper over \link{pipmd_mld}, which should be viewed
for more detail.
}
