% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_mld}
\alias{get_mld}
\title{Mean Log Deviation}
\usage{
get_mld(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to methods}
}
\value{
list
}
\description{
This is a generic function to compute the
Mean Log Deviation (MLD).
}
