% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_polarization.pipster_gd}
\alias{get_polarization.pipster_gd}
\title{Wolfson polarization index}
\usage{
\method{get_polarization}{pipster_gd}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to pipgd_polarization}
}
\value{
list
}
\description{
This is a method to compute the Wolfson polarization index on group data.
Requires a pipster object of class \code{pipster_gd} (grouped data object).
This is a wrapper over pipgd_polarization, which should be viewed
for more detail.
}
