% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_polarization.pipster_md}
\alias{get_polarization.pipster_md}
\title{Wolfson polarization index}
\usage{
\method{get_polarization}{pipster_md}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments passed to \link{pipmd_polarization}}
}
\value{
list
}
\description{
This is a method to compute the Wolfson polarization index on micro data.
Requires a pipster object of class \code{pipster_md} (micro data object).
This is a wrapper over \link{pipmd_polarization}, which should be viewed
for more detail.
}
