% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pov.R
\name{get_pov_gap.default}
\alias{get_pov_gap.default}
\title{Calculate poverty gap default method}
\usage{
\method{get_pov_gap}{default}(pipster_object, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{...}{additional arguments}
}
\value{
list
}
\description{
This default S3 method returns an error when called
}
