% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pov.R
\name{get_pov_gap.pipster_gd}
\alias{get_pov_gap.pipster_gd}
\title{Calculate poverty gap for group data}
\usage{
\method{get_pov_gap}{pipster_gd}(pipster_object, povline, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{...}{additional arguments passed to \link{pipgd_pov_gap}}
}
\value{
list
}
\description{
This is a method to compute poverty gap on group data.
Requires a pipster object of class \code{pipster_gd} (grouped data object).
This is a wrapper over \link{pipgd_pov_gap}, which should be viewed
for more detail.
}
