% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pov.R
\name{get_pov_severity.pipster_md}
\alias{get_pov_severity.pipster_md}
\title{Calculate poverty severity for micro data}
\usage{
\method{get_pov_severity}{pipster_md}(pipster_object, povline, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{...}{additional arguments passed to \link{pipmd_pov_severity}}
}
\value{
list
}
\description{
This is a method to compute poverty severity on micro data.
Requires a pipster object of class \code{pipster_md} (micro data object).
This is a wrapper over \link{pipmd_pov_severity}, which should be viewed
for more detail.
}
