% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pov.R
\name{get_pov_severity}
\alias{get_pov_severity}
\title{Calculate poverty severity}
\usage{
get_pov_severity(pipster_object, povline, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}
}
\value{
list
}
\description{
This is a generic function to compute the
poverty severity - i.e. FGT 2.
}
