% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_quantile.pipster_md}
\alias{get_quantile.pipster_md}
\title{Quantile welfare values}
\usage{
\method{get_quantile}{pipster_md}(
  pipster_object,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  ...
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}

\item{...}{additional arguments passed to \link{pipmd_quantile}}
}
\value{
list
}
\description{
Gives the \code{n} quantile welfare values for the given welfare and weight vectors.
This is for \code{pipster_gd} grouped data objects. This is a wrapper
over \link{pipmd_quantile}, which should be viewed for more detail.
}
