% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_quantile_welfare_share.default}
\alias{get_quantile_welfare_share.default}
\title{Quantile welfare values}
\usage{
\method{get_quantile_welfare_share}{default}(
  pipster_object,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  ...
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}

\item{...}{additional arguments}
}
\value{
list
}
\description{
This default S3 method returns an error when called
}
