% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_quantile_welfare_share.pipster_md}
\alias{get_quantile_welfare_share.pipster_md}
\title{Quantile welfare share in micro data}
\usage{
\method{get_quantile_welfare_share}{pipster_md}(
  pipster_object,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  ...
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}

\item{...}{additional arguments passed to \link{pipmd_quantile_welfare_share}}
}
\value{
list
}
\description{
This is a generic function.
\code{get_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \link{get_welfare_share_at} gets the share of
welfare held by a particular share of the population, which is essentially
the cumulative share. Instead, \code{get_quantile_welfare_share} returns
the proportion of welfare that only the specified quantile holds.
}
\details{
Requires a pipster object of class \code{pipster_md} (micro data object).
This is a wrapper over \link{pipmd_quantile_welfare_share}, which should be viewed
for more detail.
}
