% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_quantile_welfare_share}
\alias{get_quantile_welfare_share}
\title{Quantile welfare share}
\usage{
get_quantile_welfare_share(
  pipster_object,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  ...
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}
}
\value{
list
}
\description{
This is a generic function.
\code{get_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \link{get_welfare_share_at} gets the share of
welfare held by a particular share of the population, which is essentially
the cumulative share. Instead, \code{get_quantile_welfare_share} returns
the proportion of welfare that only the specified quantile holds.
}
