% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pov.R
\name{get_watts}
\alias{get_watts}
\title{Calculate Watts poverty index}
\usage{
get_watts(pipster_object, povline, ...)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{...}{additional arguments passed to methods}
}
\value{
list
}
\description{
This is a generic function to compute the
Watts poverty index
}
