% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_welfare_share_at.pipster_gd}
\alias{get_welfare_share_at.pipster_gd}
\title{Welfare share by quantile in group data}
\usage{
\method{get_welfare_share_at}{pipster_gd}(
  pipster_object,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  ...
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}

\item{...}{additional arguments passed to \link{pipgd_welfare_share_at}}
}
\value{
list
}
\description{
Returns the share of welfare held by the specified
share of the population in the parameter \code{popshare}. Alternatively, you can
select the number of quantiles (10 be default), to estimate the corresponding
share of welfare in each.
}
\details{
Requires a pipster object of class \code{pipster_gd} (grouped data object).
This is a wrapper over \link{pipgd_welfare_share_at}, which should be viewed
for more detail.
}
