% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_pip_type.R
\name{identify_pip_type}
\alias{identify_pip_type}
\title{Identify PIP type}
\usage{
identify_pip_type(
  welfare,
  weight = rep(1, length(welfare)),
  imputation_id = NULL,
  groupdata_threshold = 200,
  verbose = getOption("pipster.verbose")
)
}
\arguments{
\item{welfare}{numeric: welfare variable, either income of consumption}

\item{weight}{numeric: expansion sample weiths. Default is a vector o 1s of
the length of \code{welfare}}

\item{imputation_id}{numeric: vector that identifies different imputations.
Default is NULL}

\item{groupdata_threshold}{numeric: threshold to discriminate between micro
data an group data. Default is 200 observations}

\item{verbose}{logical: Whether to display important messages about your data
or query}
}
\value{
character of length 1.
}
\description{
Based on welfare and weight vectors it identifies whether the data is
microdata, group data or imputed data
}
