% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_gd}
\alias{pip_gd}
\title{pip_gd}
\format{
A grouped data frame with five variables:
\describe{
\item{\code{W}}{Weights, share of population, sum up to 100}
\item{\code{X}}{welfare vector with mean welfare by decile}
\item{\code{P}}{Cumulative share of population}
\item{\code{L}}{Cumulative share of welfare}
\item{\code{R}}{share of welfare, sum up to 1}
}
}
\source{
Datt, Gaurav. 1998. “Computational Tools for Poverty Measurement and Analysis.” \url{http://ebrary.ifpri.org/utils/getfile/collection/p15738coll2/id/125673/filename/125704.pdf}, downloaded 2023-07-18

For further details, see \url{https://datanalytics.worldbank.org/PIP-Methodology/welfareaggregate.html#tgd}
}
\usage{
pip_gd
}
\description{
PIP Group Data example .
}
\keyword{datasets}
