% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_id}
\alias{pip_id}
\title{pip_id}
\format{
An imputed data frame with three variables:
\describe{
\item{\code{welfare}}{welfare (income or consumption)}
\item{\code{weight}}{population weights}
\item{\code{imputation_id}}{Imputation ID}
}
}
\usage{
pip_id
}
\description{
PIP imputed data example.
}
\keyword{datasets}
