% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_md}
\alias{pip_md}
\title{pip_md}
\format{
A microdata frame with two variables:
\describe{
\item{\code{welfare}}{welfare (income or consumption)}
\item{\code{weight}}{population weights}
}
}
\usage{
pip_md
}
\description{
PIP microdata example.
}
\keyword{datasets}
