% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_md_s}
\alias{pip_md_s}
\title{pip_md_s}
\format{
A 100 obs microdata frame with two variables:
\describe{
\item{\code{welfare}}{welfare (income or consumption)}
\item{\code{weight}}{population weights}
}
}
\usage{
pip_md_s
}
\description{
Small PIP microdata example.
}
\keyword{datasets}
