% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_params.R
\name{pipgd_params}
\alias{pipgd_params}
\title{Get Lorenz Parameters}
\usage{
pipgd_params(welfare, population)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}
}
\value{
list with Group data parameters parameters
}
\description{
Get Parameters and key values derived from the quadratic and Beta Lorenz
parametrization. \code{welfare} and \code{population} must be vectors of a group data
dataset
}
\examples{
# Get Lorenz parameters
res <- pipgd_params(
  welfare = pip_gd$L,
  population = pip_gd$P)
str(res)
}
