% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_pov.R
\name{pipgd_pov_severity_nv}
\alias{pipgd_pov_severity_nv}
\title{Estimate poverty severity (non-vectorized)}
\usage{
pipgd_pov_severity_nv(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  lorenz = NULL,
  pov_gap = NULL,
  complete = getOption("pipster.return_complete")
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{pov_gap}{list: When NULL (default), the welfare and weight
arguments are used to estimate all underlying parameters.
Else, should be a list of output from \link{pipgd_pov_gap_nv}.}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
list: contains numeric poverty severity and, if \code{complete=TRUE},
also returns all params.
}
\description{
This function is not vectorized and thus is not exported. Use
\link{pipgd_pov_severity} instead.
}
\keyword{internal}
