% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{pipgd_quantile_welfare_share}
\alias{pipgd_quantile_welfare_share}
\title{Quantile welfare share}
\usage{
pipgd_quantile_welfare_share(
  params = NULL,
  weight = NULL,
  welfare = NULL,
  complete = getOption("pipster.return_complete"),
  lorenz = NULL,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n)
)
}
\arguments{
\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{weight}{numeric vector of cumulative share of the population}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}
}
\value{
list with welfare shares
}
\description{
\code{pipgd_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \code{pipgd_welfare_share_at} get the share of
welfare held by a particular share of the population. Instead,
\code{pipgd_quantile_welfare_share} returns the proportion of welfare that only
the specified quantile holds.
}
\examples{
pipgd_quantile_welfare_share(welfare = pip_gd$L,
                        weight = pip_gd$P,
                        complete = FALSE)
}
