% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_lorenz.R
\name{pipgd_select_lorenz}
\alias{pipgd_select_lorenz}
\title{Get selected Lorenz curve for distributional stats}
\usage{
pipgd_select_lorenz(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  complete = getOption("pipster.return_complete")
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
Returns a list of values with the best lorenz fit given the
distributional Stats.
}
\description{
Get selected Lorenz curve for distributional stats
}
\examples{
# Example 1: Directly using welfare and weight vectors
pipgd_select_lorenz(welfare = pip_gd$L,
                    weight = pip_gd$P)

# Example 2: Specifying mean and poverty line
custom_mean <- mean(pip_gd$L)
custom_povline <- 1.25
pipgd_select_lorenz(welfare = pip_gd$L,
                    weight = pip_gd$P,
                    mean = custom_mean,
                    povline = custom_povline)
rm(custom_mean, custom_povline)

# Example 5: Detailed output with complete = TRUE
pipgd_select_lorenz(welfare = pip_gd$L,
                    weight = pip_gd$P,
                    complete = TRUE)

}
