% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_lorenz.R
\name{pipgd_validate_lorenz}
\alias{pipgd_validate_lorenz}
\title{Check validity of Lorenz Curve}
\usage{
pipgd_validate_lorenz(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  complete = getOption("pipster.return_complete")
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_params()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric: welfare mean of distribution.}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
Returns a nested list of distributional validity of each Lorenz model
accessible at \verb{$gd_params$lq$validity$is_normal} for the Lorenz beta and
\verb{$gd_params$lq$validity$is_normal} for the Lorenz quadratic.
}
\description{
Check validity of Lorenz Curve
}
\examples{
# Example 1: Validate Lorenz Curves using pre-calculated params
params <- pipgd_params(welfare = pip_gd$L, weight = pip_gd$P)
pipgd_validate_lorenz(params = params)

# Example 2: Directly using welfare and weight vectors
pipgd_validate_lorenz(welfare = pip_gd$L,
                      weight = pip_gd$P)

# Example 3: Specifying mean and poverty line
pipgd_validate_lorenz(welfare = pip_gd$L,
                      weight = pip_gd$P,
                      mean = mean(pip_gd$X),
                      povline = 50)

# Example 4: Using a custom population share
pipgd_validate_lorenz(welfare = pip_gd$L,
                      weight = pip_gd$P,
                      popshare = 0.5)

}
