% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_pov.R
\name{pipgd_watts}
\alias{pipgd_watts}
\title{Estimate Watts poverty index}
\usage{
pipgd_watts(
  pipster_object = pipster_object,
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  format = c("dt", "list", "atomic"),
  lorenz = NULL,
  complete = getOption("pipster.return_complete")
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{format}{character: either "dt" for data.table, "list" or "atomic" for a
single numeric vector, whose names are corresponding selected Lorenz for
each value.  Default is "dt"}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
Returns a \code{data.table} and \code{data.frame} object with two variables:
\code{watts} and \code{lorenz}.  Check \code{format} argument to change
the output format.
If \code{complete = TRUE}, it returns a \code{pipgd_params} object with additional
details and intermediate calculations.
}
\description{
Computes Watts Index from either beta or quadratic Lorenz fit.
The first distribution-sensitive poverty measure was proposed in 1968 by Watts.
It is defined as the mean across the population of the proportionate poverty
gaps, as measured by the log of the ratio of the poverty line to income,
where the mean is formed over the whole population, counting the nonpoor as
having a zero poverty gap.
}
\examples{
# Example 1: Basic usage with the pip_gd dataset and default poverty line
pipgd_watts(welfare = pip_gd$L,
            weight  = pip_gd$P)

# Example 2: Specifying a different poverty line and output as a list
pipgd_watts(welfare = pip_gd$L,
            weight  = pip_gd$P,
            povline = 1.9,
            format  = "list")


# Example 3: Detailed output with complete = TRUE
pipgd_watts(welfare  = pip_gd$L,
            weight   = pip_gd$P,
            format   = "list",
            complete = TRUE)

# Example 4: Custom mean and times_mean with data.table format
pipgd_watts(welfare    = pip_gd$L,
            weight     = pip_gd$P,
            mean       = 109.9,
            times_mean = 1.5)

}
