% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_gini}
\alias{pipmd_gini}
\title{Gini coefficient}
\usage{
pipmd_gini(welfare = NULL, weight = NULL, format = c("dt", "list", "atomic"))
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
gini
}
\description{
Compute the Gini coefficient for microdata.
}
\details{
Given a vector of income or consumption values and their respective weights
\code{pipmd_gini()} computes the Gini coefficient for the distribution.
}
\examples{
pipmd_gini(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight)
}
