% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_polarization}
\alias{pipmd_polarization}
\title{Wolfson polarization index}
\usage{
pipmd_polarization(
  welfare = NULL,
  weight = NULL,
  gini = NULL,
  mean = NULL,
  median = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{gini}{numeric: gini coefficient. If NULL (default) then uses
\link{pipmd_gini} to calculate the gini.}

\item{mean}{numeric: weighted welfare mean. Default is NULL.}

\item{median}{numeric: weighted welfare mean. Default is NULL.}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
numeric: Wolfson polarization index, see \code{format} argument.
}
\description{
Compute the Wolfson polarization index for microdata.
}
\details{
Given a vector of income or consumption values and their respective weights
\code{pipmd_polarization()} computes the Wolfson polarization index.
}
\examples{
pipmd_polarization(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight)
}
