% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_pov.R
\name{pipmd_pov_gap_nv}
\alias{pipmd_pov_gap_nv}
\title{Calculate poverty gap using microdata}
\usage{
pipmd_pov_gap_nv(
  pipster_object = NULL,
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  povline = fmean(welfare, w = weight) * times_mean,
  times_mean = 1
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights. If NULL, a vector of 1s
is used to give equal weight to each observation.}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{times_mean}{numeric factor that multiplies the mean to create a relative poverty line. Default is 1}
}
\value{
numeric: Poverty gap
}
\description{
Non-vectorized poverty gap microdata function. Use the vectorized
function \link{pipmd_pov_gap}
}
\keyword{internal}
