% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_pov.R
\name{pipmd_pov_headcount}
\alias{pipmd_pov_headcount}
\title{Calculate poverty headcount from microdata}
\usage{
pipmd_pov_headcount(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights. If NULL, a vector of 1s
is used to give equal weight to each observation.}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{format}{atomic character vector: specifies the format of output, either
"dt", "list", or "atomic"}
}
\value{
list: contains numeric poverty headcount. See \code{format}
}
\description{
Calculate poverty headcount from microdata
}
\examples{
pipmd_pov_headcount(
welfare = pipmd_s$welfare,
weight  = pipmd_s$weight,
format  = "list"
)
}
