% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_pov.R
\name{pipmd_pov_severity}
\alias{pipmd_pov_severity}
\title{Calculate poverty severity from microdata}
\usage{
pipmd_pov_severity(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  mean = 1,
  times_mean = 1,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights. If NULL, a vector of 1s
is used to give equal weight to each observation.}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{mean}{numeric scalar of distribution mean. Default is}

\item{times_mean}{numeric factor that multiplies the mean to create a relative poverty line. Default is 1}

\item{format}{atomic character vector: specifies the format of output, either
"dt", "list", or "atomic"}
}
\value{
list: contains numeric poverty severity. See \code{format}
}
\description{
Calculate poverty severity from microdata
}
\examples{
pipmd_pov_severity(
welfare = pipmd_s$welfare,
weight  = pipmd_s$weight,
format  = "list"
)
}
