% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_pov.R
\name{pipmd_pov_severity_nv}
\alias{pipmd_pov_severity_nv}
\title{Calculate poverty severity using microdata}
\usage{
pipmd_pov_severity_nv(
  welfare = NULL,
  weight = NULL,
  povline = ifelse(!is.na(mean * times_mean), mean * times_mean, NULL),
  mean = 1,
  times_mean = 1
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights. If NULL, a vector of 1s
is used to give equal weight to each observation.}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{mean}{numeric scalar of distribution mean. Default is 1.}

\item{times_mean}{numeric factor that multiplies the mean to create a relative poverty line. Default is 1}
}
\value{
numeric: Poverty severity
}
\description{
Non-vectorized poverty severity microdata function. Use the vectorized
function \link{pipmd_pov_severity}
}
\keyword{internal}
