% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_quantile_welfare_share}
\alias{pipmd_quantile_welfare_share}
\title{Quantile welfare share}
\usage{
pipmd_quantile_welfare_share(
  welfare = NULL,
  weight = NULL,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{n}{numeric: number of equi-spaced quantiles}

\item{popshare}{numeric atomic vector: the quantiles to return. Will only be
used if \code{n = NULL}}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{pipmd_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \code{pipmd_welfare_share_at} get the share of
welfare held by a particular share of the population, which is in a sense
the cumulative share. Instead, \code{pipmd_quantile_welfare_share} returns
the proportion of welfare that only the specified quantile holds.
}
\examples{
pipmd_quantile_welfare_share(welfare = pip_md_s$welfare,
                             weight = pip_md_s$weight)
}
