% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipster_gd.R
\name{pipster_gd}
\alias{pipster_gd}
\alias{is_pipster_gd}
\alias{as_pipster_gd}
\title{\code{pipster_gd} vector}
\usage{
pipster_gd(x = double())

is_pipster_gd(x)

as_pipster_gd(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{pipster_gd}: numeric vector
\item For \code{is_pipster_gd}: object to test
\item For \code{as_pipster_gd}: object that should inherit \code{pipster_gd} class
}}
}
\value{
an S3 vector of class \code{pipster_gd}
}
\description{
This creates a double vector of group pipster data.
}
\examples{
pipster_gd(c(0.1, 0.5, 0.9, 1))
}
