% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_format_md}
\alias{return_format_md}
\title{Return data according to format - microdata}
\usage{
return_format_md(
  ld,
  var,
  povline,
  complete = FALSE,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{ld}{list of data}

\item{var}{character: name of variable to be returned.}

\item{povline}{numeric: poverty line}

\item{complete}{logical: if \code{format = "list"} then \code{complete = TRUE} gives complete
information output.}

\item{format}{character: either "dt" for data.table, "list" or "atomic" for a
single numeric vector, whose names are corresponding selected Lorenz for
each value.  Default is "dt"}
}
\value{
determined by \code{format}
}
\description{
Return data according to format - microdata
}
