% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_format_md_dist}
\alias{return_format_md_dist}
\title{return md dist data format}
\usage{
return_format_md_dist(p, name, format = "atomic")
}
\arguments{
\item{p}{object from md_dist functions}

\item{name}{character: name of the indicator}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
depending on format.
}
\description{
return md dist data format
}
\keyword{internal}
