% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_pip_type.R
\name{sum_up_to_one}
\alias{sum_up_to_one}
\title{check if elements of vector sum up to 1 or (100)}
\usage{
sum_up_to_one(x, digits = 0)
}
\arguments{
\item{x}{numeric: vector whose elements might sum up to 1 or 100}

\item{digits}{numeric: threshold of the sum. Default is 0 decimals}
}
\value{
logical
}
\description{
This functions uses a heuristic to round up the sum to the closest integers,
as it is possible to have vectors that are intended to sum up to 1, but for
they fail because of precision.
}
\keyword{internal}
