% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pip.R
\name{as_pip}
\alias{as_pip}
\alias{as_pip.data.frame}
\alias{as_pip.list}
\title{Assign PIP class to object}
\usage{
as_pip(x)

\method{as_pip}{data.frame}(x)

\method{as_pip}{list}(x)
}
\arguments{
\item{x}{dataframe or list. If dataframe, the number of unique elements in the variable \code{module} should be 1}
}
\value{
data.frame or list
}
\description{
Assign PIP class to object
}
