% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{pipgd_gini}
\alias{pipgd_gini}
\title{Compute Gini coefficient}
\usage{
pipgd_gini(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  complete = getOption("pipster.return_complete"),
  lorenz = NULL
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}
}
\value{
Returns a nested list containing distribution statistics:
\verb{$dist_stats$gini} is a numeric vector containing the gini coefficient.
\verb{$dist_stats$lorenz} is a numeric vector showing the lorenz curve used.
If \code{complete = TRUE}, it returns a \code{pipgd_params} object with additional
details and intermediate calculations.
}
\description{
Gini is computed using either the beta or quadratic Lorenz
functions.
}
\examples{
# Example 1: Basic Calculation of Gini Coefficient
pipgd_gini(welfare = pip_gd$L,
           weight = pip_gd$P)

# Example 2: Gini Coefficient with a Specific Lorenz Curve (e.g. Lorenz beta)
pipgd_gini(welfare = pip_gd$L,
           weight = pip_gd$P,
           lorenz = "lb")

# Example 3: Detailed Output of Gini Calculation
pipgd_gini(welfare = pip_gd$L,
           weight = pip_gd$P,
           complete = TRUE)

# Example 4: Gini Coefficient with Adjusted Mean
pipgd_gini(welfare = pip_gd$L,
           weight = pip_gd$P,
           times_mean = 1.5)

# Example 5: Focusing on a Subset Below Poverty Line at 50
pipgd_gini(welfare = pip_gd$L,
           weight = pip_gd$P,
           povline = 50)

}
