% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{pipgd_mld}
\alias{pipgd_mld}
\title{Compute MLD}
\usage{
pipgd_mld(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  complete = getOption("pipster.return_complete"),
  lorenz = NULL
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative poverty line. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}
}
\value{
Returns a nested list containing distribution statistics:
\verb{$dist_stats$mld} is a numeric vector containing the mld calculation.
\verb{$dist_stats$lorenz} is a numeric vector showing the lorenz curve used.
If \code{complete = TRUE}, it returns a \code{pipgd_params} object with additional
details and intermediate calculations.
}
\description{
MLD (Mean Logarithimic Deviation) is computed using either the beta or
quadratic Lorenz functions.
}
\examples{
# Example 1: Basic MLD Calculation
pipgd_mld(welfare = pip_gd$L,
          weight = pip_gd$P)

# Example 2: MLD with a Specific Lorenz Curve
pipgd_mld(welfare = pip_gd$L,
          weight = pip_gd$P,
          lorenz = "lb")

# Example 3: Detailed Output of MLD Calculation
pipgd_mld(welfare = pip_gd$L,
          weight = pip_gd$P,
          complete = TRUE)

# Example 4: Adjusting for a Specific Mean
actual_mean <- 90  # Replace with the actual mean of your data
pipgd_mld(welfare = pip_gd$L,
          weight = pip_gd$P,
          mean = actual_mean)


# Example 5: MLD Focusing on Specific Poverty Line
pipgd_mld(welfare = pip_gd$L,
          weight = pip_gd$P,
          povline = 50)

}
