% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_params.R
\name{pipgd_params}
\alias{pipgd_params}
\title{Get Group Data Parameters}
\usage{
pipgd_params(welfare, weight, mean = NULL, population = NULL)
}
\arguments{
\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is NULL}

\item{population}{numeric scalar with actual size of population. Default is NULL}
}
\value{
list with Group data parameters parameters
}
\description{
Get Parameters and key values derived from the quadratic and Beta Lorenz
parametrization. \code{welfare} and \code{population} must be vectors of a group data
dataset
}
\examples{
# Get Lorenz parameters
res <- pipgd_params(
  welfare = pip_gd$L,
  weight = pip_gd$P)
str(res)
}
\references{
Datt, G. 1998. "\href{https://ageconsearch.umn.edu/record/94862/}{Computational Tools For Poverty Measurement And Analysis}". FCND
Discussion Paper 50. World Bank, Washington, DC.

Krause, M. 2013. "\href{https://doi.org/10.1016/j.jeconom.2013.01.001}{Corrigendum to Elliptical Lorenz curves}". \emph{Journal of
Econometrics 174} (1): 44.

Villasenor, J., B. C. Arnold. 1989. "\href{https://EconPapers.repec.org/RePEc:eee:econom:v:40:y:1989:i:2:p:327-338}{Elliptical Lorenz curves}".
\emph{Journal of Econometrics 40} (2): 327-338.
}
