% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_params.R
\name{pipgd_params}
\alias{pipgd_params}
\title{Get Group Data Parameters}
\usage{
pipgd_params(welfare, weight, mean = NULL, population = NULL)
}
\arguments{
\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is NULL}

\item{population}{numeric scalar with actual size of population. Default is NULL}
}
\value{
list with Group data parameters parameters
}
\description{
Get Parameters and key values derived from the quadratic and Beta Lorenz
parametrization. \code{welfare} and \code{population} must be vectors of a group data
dataset
}
\examples{
# Get Lorenz parameters
res <- pipgd_params(
  welfare = pip_gd$L,
  weight = pip_gd$P)
str(res)
}
