% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{pipgd_quantile}
\alias{pipgd_quantile}
\title{Get quantile at specified shared of population}
\usage{
pipgd_quantile(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n),
  mean = 1,
  complete = getOption("pipster.return_complete"),
  lorenz = NULL
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}

\item{mean}{numeric: welfare mean of distribution.}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}
}
\value{
vector of quantiles
}
\description{
\code{pipgd_quantile} returns the quantile (i.e., monetary value) that corresponds
to shared of the population that lives below that threshold.
}
\details{
This is basically the inverse of estimating the poverty rate (headcount or
population share) below the poverty line. In this case, you provide the
headcount and \code{pipgs_quantile} returns the "poverty line".

The quantiles are calculated as function of the mean of the distribution
times an \code{x} factor. Basically, the quantile is \code{x} times the mean. By
default, the mean is equal to 1, which implies that, if no mean value if
provided, the return value is equal to \code{x}.

\strong{NOTE:} the outcome from \code{pipgd_quantile} is not necessarily the inverse of \link{pipgd_pov_headcount}. The reason for this ia that, \link{pipgd_pov_headcount} selects the Lorenz parametrization that fits better that the specified point of the distribution (e.i., the poverty lines). \link{pipgd_quantile}, in contrast, use the same Lorenz parametrization for any point. The lorenz used is the one that fits best for all distributional measures.
}
\examples{
pipgd_quantile(welfare = pip_gd$L,
               weight  = pip_gd$P)
}
