% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_lorenz.R
\name{pipgd_select_lorenz}
\alias{pipgd_select_lorenz}
\title{Get selected Lorenz curve for distributional stats}
\usage{
pipgd_select_lorenz(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NA_real_),
  complete = getOption("pipster.return_complete")
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{mean}{numeric scalar of distribution mean. Default is 1}

\item{times_mean}{numeric factor that multiplies the mean to create a
relative povertyline. Default is 1}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
list of values with best lorenz fit for distributional Stats
}
\description{
Get selected Lorenz curve for distributional stats
}
\examples{
# Using Lorenz parameters from get_gd_lorenz_params
withr::local_options(pipster.return_complete  = TRUE)
params <- pipgd_validate_lorenz(
  welfare = pip_gd$L,
  weight = pip_gd$P)

params <- pipgd_validate_lorenz(
  params = params,
  complete = TRUE)
pipgd_select_lorenz(params = params)

# Using Lorenz parameters from pipgd_validate_lorenz
params <- pipgd_validate_lorenz(
  welfare = pip_gd$L,
  weight = pip_gd$P,
  complete = TRUE)
pipgd_select_lorenz(params = params)

# Using original vectors

pipgd_select_lorenz(
  welfare = pip_gd$L,
  weight = pip_gd$P)
}
