% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{pipgd_welfare_share_at}
\alias{pipgd_welfare_share_at}
\title{Welfare share by quantile in group data}
\usage{
pipgd_welfare_share_at(
  params = NULL,
  welfare = NULL,
  weight = NULL,
  complete = getOption("pipster.return_complete"),
  lorenz = NULL,
  n = 10,
  popshare = seq(from = 1/n, to = 1, by = 1/n)
)
}
\arguments{
\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{n}{numeric scalar for the number of quantiles to be used in \code{popshare}}

\item{popshare}{numeric: vector of share of population. Default is \code{seq(from = 1/n, to = 1, by = 1/n)}}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{pipgd_welfare_share_at} returns the share of welfare held by the specified
share of the population in the parameter \code{popshare}. Alternatively, you can
select the number of quantiles (10 be default), to estimate the corresponding
share of welfare in each.
}
\examples{
pipgd_welfare_share_at(welfare = pip_gd$L,
                        weight = pip_gd$P,
                        complete = FALSE)
}
