% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_gini}
\alias{pipmd_gini}
\title{Gini coefficient}
\usage{
pipmd_gini(
  pipster_object = NULL,
  welfare,
  weight = rep(1, length = length(welfare)),
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
Returns a \code{data.table} and \code{data.frame} object with two variables:
\code{indicator} and \code{value}. Check \code{format} argument to change the output format.
}
\description{
Compute the Gini coefficient for microdata.
}
\details{
Given a vector of income or consumption values and their respective weights
\code{pipmd_gini()} computes the Gini coefficient for the distribution.
}
\examples{
# Example 1: Basic Gini coefficient calculation with default settings
pipmd_gini(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight)

# Example 2: Returning Gini coefficient in data.table format
pipmd_gini(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight,
           format = "dt")

# Example 3: Returning Gini coefficient as an atomic value
pipmd_gini(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight,
           format = "atomic")

# Example 4: Using equal weights (weight vector is NULL)
pipmd_gini(welfare = pip_md_s$welfare,
           format = "list")

}
