% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_mld}
\alias{pipmd_mld}
\title{Mean Log Deviation}
\usage{
pipmd_mld(
  welfare = NULL,
  weight = NULL,
  mean = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{mean}{numeric: weighted welfare mean. Default is NULL.}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
Returns a \code{data.table} and \code{data.frame} object with two variables:
\code{indicator} and \code{value}. Check \code{format} argument to change the output format.
}
\description{
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD).
}
\examples{
# Example 1: Basic usage with default settings
pipmd_mld(welfare = pip_md_s$welfare,
          weight = pip_md_s$weight)

# Example 2: Specifying mean and returning as a list
custom_mean <- mean(pip_md_s$welfare)
pipmd_mld(welfare = pip_md_s$welfare,
          weight = pip_md_s$weight,
          mean = custom_mean,
          format = "list")

# Example 3: Returning MLD as an atomic value
pipmd_mld(welfare = pip_md_s$welfare,
          weight = pip_md_s$weight,
          format = "atomic")

# Example 4: Using equal weights (weight vector is NULL)
pipmd_mld(welfare = pip_md_s$welfare,
          format = "list")

}
