% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_dist.R
\name{pipmd_mld}
\alias{pipmd_mld}
\title{Mean Log Deviation}
\usage{
pipmd_mld(
  welfare = NULL,
  weight = NULL,
  mean = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector}

\item{mean}{numeric: weighted welfare mean. Default is NULL.}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
numeric: MLD, see \code{format} argument.
}
\description{
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD).
}
\examples{
pipmd_mld(welfare = pip_md_s$welfare,
           weight = pip_md_s$weight)
}
