% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipmd_pov.R
\name{pipmd_pov_severity}
\alias{pipmd_pov_severity}
\title{Calculate poverty severity from microdata}
\usage{
pipmd_pov_severity(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  mean = 1,
  times_mean = 1,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights. If NULL, a vector of 1s
is used to give equal weight to each observation.}

\item{povline}{numeric: Poverty line in international dollars, same units as welfare.}

\item{mean}{numeric scalar of distribution mean. Default is 1.}

\item{times_mean}{numeric factor that multiplies the mean to create a relative poverty line. Default is 1}

\item{format}{atomic character vector: specifies the format of output, either
"dt", "list", or "atomic"}
}
\value{
A \code{data.table} and \code{data.frame} object of length equal to the povline
vector with variables \code{povline} and \code{pov_severity}.
See \code{format} to change the output format.
}
\description{
Calculate poverty severity from microdata
}
\examples{
# Example 1: Basic poverty headcount calculation
pipmd_pov_severity(welfare = pip_md_s$welfare,
                   weight  = pip_md_s$weight,
                   povline = 1.3,
                   format  = "list")

# Example 2: Returning data.table format and multiple povline
pipmd_pov_severity(welfare = pip_md_s$welfare,
                   weight  = pip_md_s$weight,
                   povline = c(1.3, 1.2),
                   format  = "dt")

# Example 3: Returning atomic format
pipmd_pov_severity(welfare = pip_md_s$welfare,
                   weight  = pip_md_s$weight,
                   povline = 1.3,
                   format  = "atomic")

}
