% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipster.R
\docType{package}
\name{pipster}
\alias{pipster}
\alias{.datatable.aware}
\title{pipster: Poverty and Inequality methodology of WB methodology}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
A higher-level package to estimate Socioeconomic indicators on poverty and
inequality using the methodology by the WorlD Bank. This packages is mainly a
wrapper of the lower-level package \code{wbpip}
}
\section{pipster functions}{
 The pipster functions are divided in X groups.
The first, and most important is a set of functions to estimate especif
indicators using as input the welfare and weights vector of microdata or
group data.
}

\keyword{datasets}
