% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipster.R
\docType{package}
\name{pipster}
\alias{pipster}
\title{pipster: Poverty and Inequality methodology of WB methodology}
\description{
A higher-level package to estimate socioeconomic indicators on poverty and
inequality using the methodology by the World Bank. This packages is mainly a
wrapper of the lower-level package \code{wbpip}
}
\section{pipster functions}{
 The pipster functions are divided in X groups.
The first, and most important is a set of functions to estimate especif
indicators using as input the welfare and weights vector of microdata or
group data.
}

