% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipster_md.R
\name{pipster_md}
\alias{pipster_md}
\alias{is_pipster_md}
\alias{as_percent}
\title{\code{pipster_md} vector}
\usage{
pipster_md(x = double())

is_pipster_md(x)

as_percent(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{pipster_md}: numeric vector
\item For \code{is_pipster_md}: object to test
\item For \code{as_pipster_md}: object that should inherit \code{pipster_md} class
}}
}
\value{
an S3 vector of class \code{pipster_md}
}
\description{
This creates a double vector of group pipster data.
}
\examples{
pipster_md(c(0.1, 0.5, 0.9, 1))
}
