% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipgd_dist.R
\name{validate_params}
\alias{validate_params}
\title{Validate group data parameters}
\usage{
validate_params(pipster_object, welfare, weight, params)
}
\arguments{
\item{pipster_object}{pipster object created using \link{create_pipster_object}}

\item{welfare}{numeric vector of cumulative share of welfare (income/consumption)}

\item{weight}{numeric vector of cumulative share of the population}

\item{params}{list of parameters from \code{pipgd_validate_lorenz()}}
}
\value{
list: \code{params} to be used in gd functions
}
\description{
Validate group data parameters
}
\keyword{internal}
